09-deep-researcher/requirements.txt
fastapi
uvicorn
pydantic
httpx
readability-lxml
beautifulsoup4
markdownify
